backend/requirements.txt
-i https://pypi.org/simple
alembic==1.8.1
aniso8601==9.0.1
apscheduler==3.9.1
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
flask==2.1.3
flask-cors==3.0.10
flask-migrate==3.1.0
flask-restful==0.3.9
flask-sqlalchemy==2.5.1
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.1
markupsafe==2.1.1
marshmallow==3.17.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-mock==3.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
setuptools==63.2.0
six==1.16.0
sqlalchemy==1.4.39
tomli==2.0.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
werkzeug==2.1.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
